+++ satella/instrumentation/metrics/structures/__init__.py
@@ -1,1 +1,2 @@
 from .threadpool import MetrifiedThreadPoolExecutor
+from .cache_dict import MetrifiedCacheDict