+++ tests/test_instrumentation/test_metrics/test_structures.py
@@ -3,7 +3,8 @@ import unittest
 from satella.instrumentation.metrics import getMetric
 
 import time
-from satella.instrumentation.metrics.structures import MetrifiedThreadPoolExecutor
+from satella.instrumentation.metrics.structures import MetrifiedThreadPoolExecutor, \
+    MetrifiedCacheDict
 from .test_metrics import choose
 
 
@@ -12,6 +13,27 @@ def wait():
 
 
 class TestThreadPoolExecutor(unittest.TestCase):
+    def test_metrified_cache_dict(self):
+        cache_hits = getMetric('cachedict.hits', 'cps', time_unit_vectors=[10])
+        cache_miss = getMetric('cachedict.miss', 'cps', time_unit_vectors=[10])
+        value = 2
+
+        def getter(key):
+            nonlocal value
+            if value is None:
+                raise KeyError()
+            else:
+                return value
+
+        mcd = MetrifiedCacheDict(1, 2, getter, cache_hits_cps=cache_hits,
+                                 cache_miss_cps=cache_miss)
+        mcd[2]
+        self.assertEqual(next(iter(cache_hits.to_metric_data().values)).value, 0)
+        self.assertEqual(next(iter(cache_miss.to_metric_data().values)).value, 1)
+        mcd[2]
+        self.assertEqual(next(iter(cache_hits.to_metric_data().values)).value, 1)
+        self.assertEqual(next(iter(cache_miss.to_metric_data().values)).value, 1)
+
     def test_metrified_thread_pool_executor(self):
         waiting_summary = getMetric('mtpe.summary', 'summary')
         executing_summary = getMetric('mtpe.executing', 'summary')