+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.9.1
 
 * `CacheDict`: `try_delete` -> `invalidate`
-
+* added `default_value_factory` to `CacheDict`