+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * `CacheDict`: `try_delete` -> `invalidate`
 * added `default_value_factory` to `CacheDict`
 * added `unique`
+* `MetrifiedCacheDict`