+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.9.2
+
+* add extra metrics to `MetrifiedCacheDict`
+