+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.9.2
 
 * refactored metrics in `MetrifiedCacheDict`
+* cps is pending deprecated